normed functions + files in parsing
