+++ parsing/check_quotes.c
@@ -12,6 +12,50 @@
 
 #include "../minishell.h"
 
+void	init_quote_struct(t_quote *quote)
+{
+	quote->lock = 0;
+	quote->simpleq = false;
+	quote->doubleq = false;
+}
+
+int	open_or_close(t_quote *quote)
+{
+	if (quote->simpleq || quote->doubleq)
+		return (0);
+	return (1);
+}
+
+int	check_unclosed_quote(char *str)
+{
+	int i;
+	t_quote quote;
+
+	i = -1;
+	init_quote_struct(&quote);
+	while (str[++i])
+	{
+		if (quote.lock == 0 && (str[i] == '\'' || str[i] == '\"') && (str[i-1] && str[i-1] != '\\'))
+		{
+			if (str[i] == '\'')
+				quote.simpleq = !quote.simpleq;
+			else if (str[i] == '\"')
+				quote.doubleq = !quote.doubleq;
+			quote.lock = 1;
+		}
+		else if (quote.lock == 1 && ((str[i] == '\'' && quote.simpleq ) || (str[i] == '\"' && quote.doubleq)) && (str[i-1] && str[i-1] != '\\'))
+		{
+			if (quote.simpleq && str[i] == '\'')
+				quote.simpleq = !quote.simpleq;
+			else if (quote.doubleq && str[i] == '\"')
+				quote.doubleq = !quote.doubleq;
+			quote.lock = 0;
+		}
+	}
+	return (open_or_close(&quote));
+}
+
+/*
 int	check_quotes(char *str)
 {
 	int i;
@@ -35,4 +79,4 @@ int	check_quotes(char *str)
 		return (1);
 	else
 		return (0);
-}
+}*/
